+++ gulpfile.js
@@ -111,8 +111,8 @@ task('json-schema', function() {
   const inPath = 'src/analysis-format.ts';
   const outPath = 'lib/analysis.schema.json';
   return gulp.src(inPath).pipe(newer(outPath)).pipe(shell([
-    `./node_modules/.bin/typescript-json-schema --required ${inPath
-    } Analysis > ${outPath}.temp`,
+    `sh -c "./node_modules/.bin/typescript-json-schema --required ${inPath
+    } Analysis > ${outPath}.temp"`,
     `mv ${outPath}.temp ${outPath}`
   ]));
 });